ref: Rename to just /stats.
